refactor(test): P5-4 final cut — extract lifecycle-command shell-out tests
Phase 5 P5-4 final follow-up. CLI lifecycle-command shell-out tests
extracted from test_cli_shellout.cpp into test_cli_shellout_lifecycle.cpp
(14 TEST_CASEs, 14 assertions, 415 lines).

Covers the lifecycle / diagnostics surface of the CLI:
  * pulp doctor (android|ios subcommands, --validators, --versions,
    --au-cache --dry-run)
  * pulp dev (SDK-ahead-of-CLI fail-fast, value option validation)
  * pulp design (value option validation before autobind)
  * pulp sdk install (parser validation before side effects)
  * pulp upgrade (--notes from/to, --notes --json, parser errors,
    --check-only disabled, --notes no-hop, --notes --json
    slash-command-parseable)

test_cli_shellout.cpp: 2,314 → 1,919 (-395 lines, **under the
< 2,000-line P5-4 target** for the first time since the 3,439-line
baseline). All P5 cleanup targets now hit.
